Refactor to extract JWT function
